+++ src/Jih.hs
@@ -1,0 +1,3 @@
+module Jih (jihCLI) where
+
+import Jih.CLI