+++ src/Jira/Aliases.hs
@@ -1,5 +1,8 @@
 module Jira.Aliases ( AccountType
+                    , Field
+                    , FromTo
                     , IssueKey
+                    , JQLQuery
                     , Uri
                     ) where
 
@@ -9,6 +12,12 @@ import Data.Text
 -- account type is one of: 'atlassian' (normal user), 'app' (application user) or
 -- 'customer' (Jira Service Desk customer user)
 type AccountType = Text
+
+type FromTo = (Maybe Text, Maybe Text)
+
+type Field = Text
+
+type JQLQuery = Text
 
 type IssueKey = Text