+++ ChangeLog.md
@@ -2,6 +2,19 @@
 
 Format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [0.0.2] - 2021-09-06
+
+### Added
+
+- Support for a YAML config file (`~/.atlassian.yaml` by default)
+- Get issues list to process via a `--jql <JQL query>`
+- Option to display version
+
+### Changed
+
+- Need to specify `--issues` flag to process issues given as arguments
+
+
 ## [0.0.1] - 2021-08-15
 
 ### Added